+++ app/components/Views/FAQ/FAQAnswer/index.js
@@ -4,26 +4,7 @@ import { observer } from 'mobx-react';
 import OnboardingScreenWithBg from '../../../UI/OnboardingScreenWithBg';
 import { getNavigationOptionsTitle } from '../../../UI/Navbar';
 import { strings } from '../../../../../locales/i18n';
-import { RFValue } from 'react-native-responsive-fontsize';
-import { colors, fontStyles } from '../../../../styles/common';
-
-const styles = StyleSheet.create({
-    wrapper: {
-        padding: 15,
-    },
-    question: {
-        fontSize: RFValue(24),
-        ...fontStyles.bold,
-        color: colors.white,
-        marginBottom: 15,
-    },
-    answer: {
-        fontSize: RFValue(14),
-        color: colors.white,
-        ...fontStyles.normal,
-        lineHeight: 24
-    }
-});
+import styles from './styles';
 
 class FAQAnswer extends PureComponent {
     static navigationOptions = ({ navigation }) => getNavigationOptionsTitle(strings('drawer.faq_answer'), navigation);
