+++ app/components/Views/FAQ/index.js
@@ -1,33 +1,13 @@
 import React, { PureComponent } from 'react';
-import { ActivityIndicator, ScrollView, StyleSheet, View, Text } from 'react-native';
+import { ActivityIndicator, ScrollView, View } from 'react-native';
 import { observer } from 'mobx-react';
 import OnboardingScreenWithBg from '../../UI/OnboardingScreenWithBg';
 import { getNavigationOptionsTitle } from '../../UI/Navbar';
 import { strings } from '../../../../locales/i18n';
 import SettingsDrawer from '../../UI/SettingsDrawer';
-import { colors } from '../../../styles/common';
 import { makeObservable, observable } from 'mobx';
 import APIService from '../../../services/APIService';
-
-const styles = StyleSheet.create({
-	wrapper: {
-		backgroundColor: colors.transparent,
-		flex: 1,
-		paddingHorizontal: 15,
-		zIndex: 99999999999999
-	},
-	loading: {
-		position: 'absolute',
-		left: 0,
-		right: 0,
-		top: 0,
-		bottom: 0,
-		alignItems: 'center',
-		justifyContent: 'center'
-	}
-});
-
-
+import styles from "./styles";
 class FAQScreen extends PureComponent {
 	static navigationOptions = ({ navigation }) => getNavigationOptionsTitle(strings('drawer.faq'), navigation);
